Add files via upload
This commit is an implementation of Jquery-UI Sortable which can be studied here http://api.jqueryui.com/sortable/
The jQuery UI Sortable plugin makes selected elements sortable by dragging with the mouse. Here we used the same for implementing reordering of tabs and correspondingly showing content of selected tabs just like in a browser.